Move hw_mod to own file.
